+++ wasm/ts/types.ts
@@ -69,6 +69,7 @@ export interface Graph<T extends NodeId> {
   removeEdge(source: T, target: T): void;
   getEdgeWeight(source: T, target: T): number;
   setEdgeWeight(source: T, target: T, weight: number): void;
+  subgraph(nodes: T[]): Graph<T>;
   hasNodeAttr(id: T, key: string): boolean;
   getNodeAttr(id: T, key: string): AttributeValue;
   tryGetNodeAttr(id: T, key: string): AttributeValue | null;
@@ -109,9 +110,12 @@ export interface Graph<T extends NodeId> {
   dispose(): void;
 }
 
-export interface DiGraph<T extends NodeId> extends Graph<T> {}
+export interface DiGraph<T extends NodeId> extends Graph<T> {
+  subgraph(nodes: T[]): DiGraph<T>;
+}
 
 export interface MultiGraph<T extends NodeId> extends Graph<T> {
+  subgraph(nodes: T[]): MultiGraph<T>;
   hasEdgeId(edgeId: number): boolean;
   edgeIds(): number[];
   edgeIdsBetween(source: T, target: T): number[];
@@ -126,4 +130,6 @@ export interface MultiGraph<T extends NodeId> extends Graph<T> {
   removeEdgeById(edgeId: number): void;
 }
 
-export interface MultiDiGraph<T extends NodeId> extends MultiGraph<T> {}
+export interface MultiDiGraph<T extends NodeId> extends MultiGraph<T> {
+  subgraph(nodes: T[]): MultiDiGraph<T>;
+}